collector/requirements.txt
scrapy==2.11.2
apscheduler==3.10.4
httpx==0.27.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
playwright==1.44.0
scrapy-playwright==0.0.33
pyyaml==6.0.1
fake-useragent==1.5.1
fastapi==0.111.0
uvicorn[standard]==0.29.0
